[modify] JPA 적용

+++ src/main/kotlin/toy/ysjoo/schedule/database/MemoryDatabase.kt
@@ -1,12 +1,12 @@
 package toy.ysjoo.schedule.database
 
 interface MemoryDatabase<T> {
-    fun add(id: Int, data: T): Boolean
-    fun add(data: T, isAutoIndex: Boolean = true): Int
-    fun delete(id: Int): Boolean
-    fun update(id: Int, data: T): Boolean
+    fun add(id: Long, data: T): Boolean
+    fun add(data: T, isAutoIndex: Boolean = true): Long
+    fun delete(id: Long): Boolean
+    fun update(id: Long, data: T): Boolean
     fun update(data: T): Boolean
-    fun search(id: Int): T?
-    fun search(data: T): Int?
+    fun search(id: Long): T?
+    fun search(data: T): Long?
     fun searchAll(): List<T>
 }
